refactor: extract UrlMatcher to shared module
